refactor(Screen): main 에 구현했던 기능 분리
