--- src/main/board/squares.py
@@ -1,17 +1,0 @@
-
-
-
-class Node:
-    def __init__(self, row, col, board):
-        self.row = row
-        self.col = col
-        cboard = board.gameboard
-        
-        self.occupied = None
-
-        
-        
-
-
-
-
